Move createMongoParams() to utils
This way it can be used in other scripts.

Issue: S3UTILS-76
